Update td.cu
Particle playground defined.

The "particle playground" is a volume, larger than the grid's, in which particles are allowed to remain active. As soon as a particle gets out of its playground, it become inactive, and it remains so until its deletion. Inactive particles do not contribute to counts (active particles _may_ contribute to counts: only, if they stay in-grid).